requirements.txt
amqp==5.1.1
billiard==3.6.4.0
cached-property==1.5.2
celery==5.2.6
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
dnspython==2.2.1
email-validator==1.1.3
Flask==2.1.1
Flask-Cors==3.0.10
Flask-HTTPAuth==4.5.0
Flask-JWT==0.3.2
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.0.1
Jinja2==3.1.1
jsons==1.6.1
kombu==5.2.4
MarkupSafe==2.1.1
marshmallow==3.15.0
packaging==21.3
passlib==1.7.4
prompt-toolkit==3.0.29
psycopg2-binary==2.9.3
PyJWT==1.4.2
pyparsing==3.0.8
pytz==2022.1
requests==2.27.1
simplejson==3.17.6
six==1.16.0
SQLAlchemy==1.4.35
typing-extensions==4.1.1
typish==1.9.3
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
Werkzeug==2.1.1
zipp==3.8.0
